fix(app): restore SetupWizard intro page and auth-aware key generation
Restore the "create your delegate key" intro page at /dashboard for
users who connected but have no delegate key yet. Both Enoki and
Wallet flows now see the intro before key generation.

After clicking "generate delegate key":
- Enoki: silent key gen + on-chain registration with contextual
  status text ("this may take a few seconds...")
- Wallet: show key + copy + confirm + on-chain registration with
  wallet approval prompt

Simplify LandingPage back to login popover + redirect to /dashboard.
