+++ apps/app/src/App.tsx
@@ -21,6 +21,7 @@ import { config } from './config'
 
 import LandingPage from './pages/LandingPage'
 import Dashboard from './pages/Dashboard'
+import SetupWizard from './pages/SetupWizard'
 import Playground from './pages/Playground'
 
 
@@ -133,7 +134,7 @@ function AppContent() {
       <Route path="/" element={<LandingPage />} />
       <Route path="/dashboard" element={
         !currentAccount ? <Navigate to="/" replace /> :
-        delegateKey ? <Dashboard /> : <Navigate to="/" replace />
+        delegateKey ? <Dashboard /> : <SetupWizard />
       } />
       <Route path="/playground" element={
         !currentAccount ? <Navigate to="/" replace /> :
